+++ tests/testthat/helper.R
@@ -96,13 +96,13 @@ athena_test_req3 <- list(OutputLocation = Sys.getenv("noctua_s3_query"))
 athena_test_req4 <- list(OutputLocation = Sys.getenv("noctua_s3_query"))
 
 # static s3 path location
-s3_loc <- list(exp_s3_1 = "path/to/file/test/dummy_file/dummy_file.csv",
-               exp_s3_2 = "path/to/file/YEAR=2000/dummy_file.csv.gz",
-               exp_s3_3 = c("path/to/test/dummy_file/YEAR=2000/dummy_file_1.tsv", "path/to/test/dummy_file/YEAR=2000/dummy_file_2.tsv"),
-               exp_s3_4 = c("path/to/test/dummy_file/YEAR=2000/dummy_file_1.tsv.gz","path/to/test/dummy_file/YEAR=2000/dummy_file_2.tsv.gz"),
-               exp_s3_5 = "path/to/test/dummy_file/dummy_file.parquet",
-               exp_s3_6 = "path/to/dummy_file/YEAR=2000/dummy_file.snappy.parquet",
-               exp_s3_7 = "path/to/test/dummy_file/dummy_file.json",
-               exp_s3_8 = "path/to/dummy_file/YEAR=2000/dummy_file.json")
+s3_loc <- list(exp_s3_1 = "path/to/file/test/dummy_file",
+               exp_s3_2 = "path/to/file/YEAR=2000",
+               exp_s3_3 = c("path/to/test/dummy_file/YEAR=2000", "path/to/test/dummy_file/YEAR=2000"),
+               exp_s3_4 = c("path/to/test/dummy_file/YEAR=2000","path/to/test/dummy_file/YEAR=2000"),
+               exp_s3_5 = "path/to/test/dummy_file",
+               exp_s3_6 = "path/to/dummy_file/YEAR=2000",
+               exp_s3_7 = "path/to/test/dummy_file",
+               exp_s3_8 = "path/to/dummy_file/YEAR=2000")
 
 show_ddl <- SQL(paste0('CREATE EXTERNAL TABLE `default.test_df`(\n  `w` timestamp, \n  `x` int, \n  `y` string, \n  `z` boolean)\nPARTITIONED BY ( \n  `timestamp` string)\nROW FORMAT DELIMITED \n  FIELDS TERMINATED BY \'\\t\' \n  LINES TERMINATED BY \'\\n\' \nSTORED AS INPUTFORMAT \n  \'org.apache.hadoop.mapred.TextInputFormat\' \nOUTPUTFORMAT \n  \'org.apache.hadoop.hive.ql.io.HiveIgnoreKeyTextOutputFormat\'\nLOCATION\n  \'' ,Sys.getenv("noctua_s3_tbl"), 'default/test_df\'\nTBLPROPERTIES (\n  \'skip.header.line.count\'=\'1\')'))