+++ NEWS.md
@@ -1,7 +1,10 @@
 # noctua 1.8.1.9000
 ## New Feature
 * Move `sql_escape_date` into `dplyr_integration.R` backend (RAthena: [# 121](https://github.com/DyfanJones/RAthena/issues/121)).
+* Allow noctua to append to a static AWS s3 location using uuid
 
+## Bug Fix:
+* parquet file.types now use parameter `use_deprecated_int96_timestamps` set to `TRUE`. This puts POSIXct data type in to `java.sql.Timestamp` compatible format, such as `yyyy-MM-dd HH:mm:ss[.f...]`. Thanks to Christian N Wolz for highlight this issue.
 
 # noctua 1.8.1
 ## Bug Fix